Remove cruft
The original repo is a generic base for people's projects, while
this one is for a highly specialized feature set and thus will be
only maintained by the relevant parties, so CoC or Slack's other
contribution related things aren't needed anymore. Additionally,
this won't be going to be put in a CI, so test config is removed
as well.